refactor(api): move route to team context
