+++ api/src/team/application/certification-center-membership/certification-center-membership.route.js
@@ -28,4 +28,28 @@ export const certificationCenterMembershipRoute = [
       tags: ['api', 'team', 'certification-center', 'members'],
     },
   },
+  {
+    method: 'PATCH',
+    path: '/api/certification-centers/{certificationCenterId}/certification-center-memberships/{id}',
+    config: {
+      validate: {
+        params: Joi.object({
+          certificationCenterId: identifiersType.certificationCenterId,
+          id: identifiersType.certificationCenterMembershipId,
+        }),
+      },
+      handler: certificationCenterMembershipController.updateFromPixCertif,
+      pre: [
+        {
+          method: securityPreHandlers.checkUserIsAdminOfCertificationCenter,
+          assign: 'hasAuthorizationToAccessAdminScope',
+        },
+      ],
+      notes: [
+        "- **Cette route est restreinte aux utilisateurs ayant les droits d'accès**\n" +
+          "- Modification du rôle d'un membre d'un centre de certification\n",
+      ],
+      tags: ['api', 'certification-center-membership'],
+    },
+  },
 ];
